código do guiao4
