+++ src/expMat.c
@@ -13,14 +13,9 @@
 #include <string.h>
 
 /** 
- * @brief A função `s_add()` soma dois números inteiros contidos na stack.
+ * @brief A função `add()` soma dois números inteiros contidos na stack.
  *        
  * Faz uso da função `pop()` para aceder aos operandos, ou seja, ao valor que se encontra no topo da stack e ao valor que se encontra abaixo deste.
- * 
- * - __Nota:__ Caso os inputs sejam uma combinação de arrays com arrays/inteiros/doubles ou de strings com strings/caracteres, a função efetua a
- * operação de concatenar arrays ou strings.
- * 
- * @param s Stack.
  */
 void s_add(STACK *s)
 {   
@@ -182,8 +177,6 @@ void s_add(STACK *s)
  *        
  * Faz uso da função `pop()` para aceder aos operandos, ou seja, ao valor que se encontra no topo da stack e ao valor que se encontra abaixo deste.
  * Assim, __x__ será o segundo valor introduzido pelo utilizador e __y__ o primeiro, pelo que fazemos __y - x__.
- * 
- * @param s Stack.
  */
 void subtract(STACK *s)
 {   
@@ -219,23 +212,13 @@ void subtract(STACK *s)
  * do array/string original, onde 'n' é o valor do segundo operando. Por exemplo, o input `$ [ 1 2 3 ] 2 *` teria como resultado: `123123`, tal
  * como `$ "abc" 2 *` teria como resultado `abcabc`.
  * 
- * @param s Stack.
  */
 void multiply(STACK *s)
 {   
     DADOS x = pop(s);
     DADOS y = pop(s);
     
-    if (x.tipo == BLOCK)
-    {
-        DADOS *var = malloc(sizeof(DADOS) * 26);
-        initialize_var(var);
-        
-        fold_array(s, x, y, var);
-
-        free(var);
-    }
-    else if (y.tipo == ARRAY)
+    if (y.tipo == ARRAY)
     {
         double *a = x.dados;
         long n = *a;
@@ -305,11 +288,6 @@ void multiply(STACK *s)
  *        
  * Faz uso da função `pop()` para aceder aos operandos, ou seja, ao valor que se encontra no topo da stack e ao valor que se encontra abaixo deste.
  * Assim, __x__ será o segundo valor introduzido pelo utilizador e __y__ o primeiro, pelo que fazemos __y - x__.
- * 
- * - __Nota:__ Caso os operandos sejam strings, a função `divide()` irá executar o operador de strings `/`, que está definido e documentado na função
- * auxiliar `slash_str()`.
- * 
- * @param s Stack.
  */
 void divide(STACK *s)
 {   
@@ -327,7 +305,9 @@ void divide(STACK *s)
         push_long(s, r);
     }
     else if (x.tipo == STRING && y.tipo == STRING)
+    {
         slash_str(s, x, y);
+    }
     else
     {
         double r = *b / *a;
@@ -343,8 +323,6 @@ void divide(STACK *s)
  *        
  * O resultado de AND é 1 apenas se os dois bits forem 1.
  * No final, o resultado obtido é colocado na stack através da função `push_long()`.
- * 
- * @param s Stack.
  */
 void bit_and(STACK *s)
 {
@@ -365,8 +343,6 @@ void bit_and(STACK *s)
  *        
  * O resultado de OR é 1 se um dos dois bits for 1.
  * No final, o resultado obtido é colocado na stack através da função `push_long()`.
- * 
- * @param s Stack.
  */
 void bit_or(STACK *s)
 {
@@ -387,8 +363,6 @@ void bit_or(STACK *s)
  *        
  * O resultado de XOR é 1 se os dois bits forem diferentes.
  * No final, o resultado obtido é colocado na stack através da função `push_long()`.
- * 
- * @param s Stack.
  */
 void bit_xor(STACK *s)
 {
@@ -412,8 +386,6 @@ void bit_xor(STACK *s)
  * 
  * - __Nota:__ No caso de o input ser um ARRAY, a função `bit_not()` coloca na stack todos os elementos do mesmo. É usada a mesma
  * função para ambas estas operações uma vez que os operadores (`~`) são idênticos.
- * 
- * @param s Stack.
  */
 void bit_not(STACK *s)
 {
@@ -428,14 +400,6 @@ void bit_not(STACK *s)
             DADOS r = a->stack[i];
             push(s, r);
         }
-    }
-    else if (x.tipo == BLOCK)
-    {
-        DADOS *var = malloc(sizeof(DADOS) * 26);
-        initialize_var(var);
-        execute_block(s, x, var);
-
-        free(var);
     }
     else                      // Operação NOT binária
     {
@@ -454,10 +418,6 @@ void bit_not(STACK *s)
  *        
  * Para tal, é utilizada a função `pop()`, retirando da stack o elemento a ser trabalhado.
  * No final, após a subtração, o resultado obtido é colocado na stack através da função `push_long()` ou `push_double()`.
- * 
- * - __Nota:__ Para arrays ou strings, a função `decr()` retira o primeiro elemento do array ou string e coloca-o na stack.
- * 
- * @param s Stack.
  */
 void decr(STACK *s)
 {
@@ -504,10 +464,6 @@ void decr(STACK *s)
  *        
  * Para tal, é utilizada a função `pop()`, retirando da stack o elemento a ser trabalhado.
  * No final, após a adição, o resultado obtido é colocado na stack através da função `push_long()` ou `push_double()`.
- * 
- * - __Nota:__ Para arrays ou strings, a função `incr()` retira o último elemento do array ou string e coloca-o na stack.
- * 
- * @param s Stack.
  */
 void incr(STACK *s)
 {
@@ -545,51 +501,26 @@ void incr(STACK *s)
  * @brief A função `mod()` dá-nos o módulo da divisão do segundo número da stack a contar do topo pelo que se encontra no topo.
  *        
  * Para esse fim usa `pop()` para obter os mesmos e `push()` para returnar o resultado da operação.
- * 
- * - __Nota:__ Quando o input é um bloco (BLOCK), realiza a operação de aplicar um bloco a um array/string, utilizando por isso
- * as funções `execute_block_array()` e `execute_block_string()`, cujo objetivo e funcionamento está documentado em stackBlocks.c.
- * 
- * @param s Stack.
  */
  void mod(STACK *s)
 {
-    DADOS x = pop(s);
+    double *ai = pop(s).dados;
+    long a = *ai;
 
-    DADOS y = pop(s);
+    double *bi = pop(s).dados;
+    long b = *bi;
 
-    if (x.tipo == BLOCK && y.tipo == ARRAY)
-    {
-        DADOS *var = malloc(sizeof(DADOS) * 26);
-        initialize_var(var);
-
-        execute_block_array(s, x, y, var);
-
-        free(var);
-    }
-    else if (x.tipo == BLOCK && y.tipo == STRING)
-        execute_block_string(s, x, y);
-    else
-    {
-        double *ai = x.dados;
-        long a = *ai;
-        double *bi = y.dados;
-        long b = *bi;
-
-        double r = b % a;
-        push_long(s, r);
-        free(x.dados);
-        free(y.dados);
-    }
+    double r = b % a;
+    push_long(s, r);
+    
+    free(ai);
+    free(bi);
 }
 
 /**
  * @brief Esta função tem como objetivo obter a exponenciação de valores da stack.
  *        
  * Este valor é obtido elevando o segundo número a contar de cima da stack por o do topo.
- * 
- * - __Nota:__ Caso os inputs sejam strings, a função `exp()` efetua a operação de procura de substrings em strings, devolvendo o índice do primeiro caracter da substring encontrada.
- * 
- * @param s Stack.
  */
  void expo(STACK *s)
 {
