+++ src/expStack.c
@@ -70,4 +70,6 @@ void ncopy(STACK *s)
     DADOS y = s->stack[(s->sp) - i];
 
     push(s, y);
+
+    free(t.dados);
 }
