+++ src/main.c
@@ -9,7 +9,7 @@
 #include "stack.h"
 
 /**
- * @brief ?
+ * @brief 
  * 
  * @param line String contendo a totalidade do input.
  * @param token String cujo objetivo é armazenar um operador/operando do input individualmente, para que seja depois tratado.
@@ -21,15 +21,12 @@ char* get_token(char* line, char token[])
     int index = 0;
     int flag = 0;
     int str_flag = 0;
-    int block_flag = 0;
-
     while (*line != '\n')
     {
         while (*line == ' ')
             line++;
 
-        while ((*line && *line != ' ' && *line != '\n' && (flag == 0 || str_flag == 0 || block_flag == 0)) || 
-                (flag > 0 || str_flag > 0 || block_flag > 0))
+        while ((*line && *line != ' ' && *line != '\n' && (flag == 0 || str_flag == 0)) || (flag > 0 || str_flag > 0))
         {   
             token[index] = *line;
             ++size;
@@ -38,12 +35,8 @@ char* get_token(char* line, char token[])
                 ++flag;
             else if (*line == '"' && str_flag == 0)
                 ++str_flag;
-            else if (*line == '{')
-                ++block_flag;
             else if (*line == '"' && str_flag > 0)
                 --str_flag;
-            else if (*line == '}')
-                -- block_flag;
             else if (*line == ']')
                 --flag;
             
@@ -69,9 +62,8 @@ char* get_token(char* line, char token[])
 int main()
 {
     STACK* s = new_stack();
-    DADOS* var = malloc(sizeof(DADOS) * 26);
+    DADOS var[26];
     initialize_var(var);
-    s->stack[0] = *var;
 
     char* line = malloc(sizeof(char) * BUFSIZ);
     char token[BUFSIZ];
