+++ src/stack.c
@@ -5,7 +5,6 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#include <string.h>
 #include "stack.h"
 
 // Declaração de nova stack
@@ -132,10 +131,7 @@ void push_char(STACK* s, char elem)
  */
 void push_string(STACK *s, char* elem)
 {   
-    char *elemP = malloc(sizeof(elem)+1);
-    strcpy(elemP, elem);
-    
-    DADOS d = {STRING, elemP};
+    DADOS d = {STRING, elem}; // Seria 'DADOS d = {STRING, elemP};' com o código acima.
     s->sp++;
     s->stack[s->sp] = d;
 }
@@ -146,16 +142,6 @@ void push_array(STACK *s, STACK elem)
     *arrayP = elem;
 
     DADOS d = {ARRAY, arrayP};
-    s->sp++;
-    s->stack[s->sp] = d;
-}
-
-void push_block(STACK* s, char* elem)
-{
-    char *elemP = malloc(sizeof(elem)+1);
-    strcpy(elemP, elem);
-    
-    DADOS d = {BLOCK, elemP};
     s->sp++;
     s->stack[s->sp] = d;
 }
@@ -193,11 +179,6 @@ void push(STACK* s, DADOS elem)
         STACK *n = elem.dados;
         push_array(s, *n);
     }
-    else if (elem.tipo == BLOCK)
-    {
-        char *n = elem.dados;
-        push_block(s, n);
-    }
 }
 
 // Função pop()
@@ -216,12 +197,7 @@ DADOS pop(STACK* s)
     return d;
 }
 
-/**
- * @brief Verifica se é necessário alocar mais memória para a stack, o que acontece quando esta atinge o limite de capacidade.
- * 
- * @param s Stack.
- * @return DADOS* Retorna o endereço do array de dados da stack.
- */
+
 DADOS* memory_checker(STACK* s)
 {
     if (s->sp == s->cap)
